requirements.txt
docker==7.1.0
pytest-testinfra==10.1.1
flake8==7.1.1
